Correct subclass constructor signature
As a subclass of ForeignKey, AddressField should implement a
constructor with the same signature as ForeignKey, such that if code
wishes to treat AddressField as an instance of a ForeignKey, it can
instantiate it that way. This is important for other libraries that
handle FKs. Even though we'll ignore the *args passed to AddressField,
we should add it to the signature to keep it consistent with ForeignKey.